requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alibi==0.9.6
annotated-types==0.7.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==23.2.0
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
cloudpickle==3.1.0
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
Deprecated==1.2.18
dill==0.3.9
dnspython==2.7.0
eventlet==0.35.1
Farama-Notifications==0.0.4
filelock==3.16.1
Flask==2.0.0
Flask-SocketIO==4.3.1
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.4.0
google-auth==2.37.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.69.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h5py==3.12.1
huggingface-hub==0.28.1
idna==3.10
imageio==2.36.1
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.5
joblib==1.4.2
keras==2.12.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
lightning==2.2.5
lightning-utilities==0.11.9
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.12
networkx==3.4.2
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.17.0
onnx2pytorch==0.5.1
opencv-python==4.9.0.80
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==10.2.0
preshed==3.0.9
propcache==0.2.1
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.5.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-engineio==3.13.2
python-socketio==4.5.1
pytorch-lightning==2.5.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
safetensors==0.5.2
scikit-image==0.22.0
scikit-learn==1.6.1
scipy==1.11.3
seaborn==0.13.2
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-lookups-data==1.0.5
srsly==2.5.1
statsmodels==0.14.4
sympy==1.13.3
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-addons==0.23.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf-keras-vis==0.8.7
tf2onnx==1.16.1
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2025.1.10
tokenizers==0.21.0
torch==2.3.0
torchinfo==1.8.0
torchmetrics==1.3.2
torchvision==0.18.0
tqdm==4.66.4
transformers==4.48.3
triton==2.3.0
typeguard==2.13.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==2.0.3
wrapt==1.14.1
yarl==1.18.3
